requirements.txt
requests
icalendar
python-dotenv